requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-redis==5.2.0
django-silk==5.0.3
djangorestframework==3.14.0
drf-yasg==1.21.5
environs==9.5.0
flake8==6.0.0
gprof2dot==2022.7.29
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
uuid==1.30
